+++ MappingStore.ps1
@@ -39,7 +39,7 @@ function Get-MappingStatusColumns {
         @{ Name = 'isReplaced';           Default = '0' },
         @{ Name = 'isMarked';             Default = '0' },
         @{ Name = 'isReviewed';           Default = '0' },
-        @{ Name = 'isGfixLogMarked';      Default = '0' }
+        @{ Name = 'ReviewComment';        Default = '' }
     )
 }